user controller

+++ backend/app/Http/Controllers/API/Branch/CourierController.php
@@ -22,6 +22,7 @@ class CourierController extends Controller
         $couriers = $couriers->unique('id');
         return response($couriers);
     }
+
     public function citycouriers()
     {
         // $courier = Auth::user()::with('city.courier')->orderBy('id', 'desc')->paginate(10);
@@ -36,16 +37,27 @@ class CourierController extends Controller
         return response($courier);
     }
 
+
+    public function store(CourierRequest $request)
+    {
+        $input = $request->validated();
+        $input['password'] = bcrypt($input['password']);
+        $courier = Courier::create($input);
+        $courier->city()->attach($request->city);
+        $courier->district()->attach($request->district);
+
+        return response()->json('success');
+    }
     //kurye şubeye aitse düzenlenebilir
     public function edit(Courier $courier)
     {
-        $courier = $this->checkBranchCourier($courier->id);  //kurye şubeninse dburdan devam
+        abort_unless(\Gate::allows('branch-own-couriers',$courier->id), 403);
         return response()->json($courier);
     }
 
     public function update(CourierRequest $request, Courier $courier)
     {
-        $courier = $this->checkBranchCourier($courier->id); //kurye şubeninse dburdan devam
+        abort_unless(\Gate::allows('branch-own-couriers',$courier->id), 403); //kurye şubeninse dburdan devam
         $input = $request->validated();
         if(!empty($input['password'])){
             $input['password'] = bcrypt($input['password']);
@@ -59,32 +71,18 @@ class CourierController extends Controller
 
     public function destroy(Courier $courier) // soft delete
     {
-        $courier = $this->checkBranchCourier($courier->id); //kurye şubeninse dburdan devam
+        abort_unless(\Gate::allows('branch-own-couriers',$courier->id), 403); //kurye şubeninse dburdan devam
         $courier->delete();
         return response()->json('success');
     }
 
 
-    public function tasks($courier) //kuryenin sorumlu olduğu illerdeki şubeler
+    public function tasks(Courier $courier) //kuryenin sorumlu olduğu illerdeki şubeler
     {
-        $this->checkBranchCourier($courier);
-        $tasks = Courier::with('task')->where('id', $courier)->orderBy('id', 'desc')->paginate(10);
+        abort_unless(\Gate::allows('branch-own-couriers',$courier->id), 403); //kurye şubeninse dburdan devam
+        $tasks = Courier::with('task')->where('id', $courier->id)->orderBy('id', 'desc')->paginate(10);
      // $tasks = $courier->task()->orderBy('id', 'desc')->paginate(10);
         return response($tasks);
-    }
-
-    public function checkBranchCourier($courierid) //kuryenin sorumlu olduğu illerdeki şubeler
-    {
-        $districts = Auth::user()->district;
-        $couriers=[];
-        foreach ($districts as $district) {
-            array_push($couriers,$district->courier);
-        }
-        $couriers=collect($couriers)->flatten();
-        $couriers = $couriers->unique('id');
-        $couriers = $couriers->pluck('id'); //sadece id leri çek
-        $courier = Courier::with('city','district')->whereIn('id', $couriers)->findOrFail($courierid);
-        return $courier;
     }
 
 }
